move ui stuff to module
